models and static otp

+++ member/views.py
@@ -18,13 +18,16 @@ class GenerateOTPView(APIView):
     permission_classes = [AllowAny]
     def post(self, request, *args, **kwargs):
         phone_number = request.data.get('phone_number')
-        if not MemberMaster.objects.filter(mobile_no=phone_number).exists():
-            return Response({'status': 400, 'message': _('Mobile no doest not exists')}, status=status.HTTP_400_BAD_REQUEST)
-        otp = OTP.objects.filter(phone_number=phone_number)
-        if otp:
-            otp.delete()
-            
-        notp = OTP.objects.create(phone_number=phone_number)
+        if not phone_number == '9415829988':
+            if not MemberMaster.objects.filter(mobile_no=phone_number).exists():
+                return Response({'status': 400, 'message': _('Mobile no doest not exists')}, status=status.HTTP_400_BAD_REQUEST)
+            otp = OTP.objects.filter(phone_number=phone_number)
+            if otp:
+                otp.delete()
+            notp = OTP.objects.create(phone_number=phone_number)
+        else:
+              notp = '112233'
+        
         send_sms_api(mobile=phone_number, otp=notp)
         return Response({'status': 200, 'message': _('OTP sent')}, status=status.HTTP_200_OK)
